# apps/ios/Packages/DenAPI/Sources/DenAPI/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Den API",
    "description": "M3 REST API. Cookie-authenticated writes require Origin and X-CSRF-Token. CLI and agents use bearer authentication. Entity IDs are ULIDs; background image IDs are opaque SHA-256 content IDs.",
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/channels/{id}/messages": {
      "get": {
        "tags": [
          "messages"
        ],
        "operationId": "get__channels__id__messages",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "roots_only",
            "in": "query",
            "description": "True returns the main conversation only. Absent or false is the legacy flat list.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              }
            }
          },
          "default": {
            "description": "API error; invalid input, authentication, permissions, conflict, throttling or storage failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "messages"
        ],
        "operationId": "post__channels__id__messages",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessage"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "default": {
            "description": "API error; invalid input, authentication, permissions, conflict, throttling or storage failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/messages/{id}": {
      "get": {
        "tags": [
          "activity"
        ],
        "operationId": "get__messages__id_",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "default": {
            "description": "API error; invalid input, authentication, permissions, conflict, throttling or storage failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "messages"
        ],
        "operationId": "delete__messages__id_",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "default": {
            "description": "API error; invalid input, authentication, permissions, conflict, throttling or storage failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "messages"
        ],
        "operationId": "patch__messages__id_",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditMessage"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "default": {
            "description": "API error; invalid input, authentication, permissions, conflict, throttling or storage failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/uploads": {
      "post": {
        "tags": [
          "uploads"
        ],
        "operationId": "post__uploads",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeginUpload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Upload"
                }
              }
            }
          },
          "413": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "default": {
            "description": "API error; invalid input, authentication, permissions, conflict, throttling or storage failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "BeginUpload": {
        "type": "object",
        "required": [
          "channel_id",
          "filename",
          "content_type",
          "size"
        ],
        "properties": {
          "channel_id": {
            "$ref": "#/components/schemas/String"
          },
          "content_type": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "CreateMessage": {
        "type": "object",
        "required": [
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "reply_to": {
            "type": [
              "string",
              "null"
            ]
          },
          "task_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Opaque job identity supplied by a runner, scoped to this author and channel."
          },
          "thread_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Explicit thread destination. Absent is not by itself a main-conversation\nsend: reply_to or task_id may still place the message in a thread."
          },
          "upload_ids": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/String"
            }
          }
        }
      },
...
      "EditMessage": {
        "type": "object",
        "required": [
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          }
        }
      },
...
      "Message": {
        "type": "object",
        "required": [
          "id",
          "channel_id",
          "author_id",
          "content",
          "created_at",
          "attachments"
        ],
        "properties": {
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Upload"
            }
          },
          "author_id": {
            "$ref": "#/components/schemas/String"
          },
          "channel_id": {
            "$ref": "#/components/schemas/String"
          },
          "content": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "edited_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "$ref": "#/components/schemas/String"
          },
          "mention_ids": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/String"
            }
          },
          "objects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObjectSummary"
            }
          },
          "reactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Reaction"
            }
          },
          "reply_to": {
            "type": [
              "string",
              "null"
            ]
          },
          "thread": {
            "type": [
              "object",
              "null"
            ],
            "description": "Shared thread metadata. Deliberately free of per-user state so one payload can\nbe broadcast to every member who can see the parent channel.",
            "required": [
              "id",
              "channel_id",
              "root_message_id",
              "title",
              "created_by",
              "created_at",
              "reply_count",
              "last_activity_at"
            ],
            "properties": {
              "channel_id": {
                "$ref": "#/components/schemas/String"
              },
              "created_at": {
                "type": "string"
              },
              "created_by": {
                "$ref": "#/components/schemas/String"
              },
              "id": {
                "$ref": "#/components/schemas/String"
              },
              "last_activity_at": {
                "type": "string"
              },
              "last_reply_id": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "reply_count": {
                "type": "integer",
                "format": "int64"
              },
              "resolved_at": {
                "type": [
                  "string",
                  "null"
                ],
                "description": "Both resolution fields are set together or not at all."
              },
              "resolved_by": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "root_message_id": {
                "$ref": "#/components/schemas/String",
                "description": "The message the conversation hangs off. It stays in the main conversation."
              },
              "title": {
                "type": "string",
                "description": "A snapshot inferred at creation. Editing the root never rewrites it."
              }
            }
          },
          "thread_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Set on replies. Roots and main-conversation messages leave it absent."
          }
        }
      },
...
      "ObjectSummary": {
        "type": "object",
        "description": "Document records keyed by record id. The server treats record contents as opaque.",
        "required": [
          "id",
          "channel_id",
          "kind",
          "name",
          "version",
          "created_by",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "channel_id": {
            "$ref": "#/components/schemas/String"
          },
          "created_at": {
            "type": "string"
          },
          "created_by": {
            "$ref": "#/components/schemas/String"
          },
          "id": {
            "$ref": "#/components/schemas/String"
          },
          "kind": {
            "type": "string"
          },
          "message_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "thumbnail_upload_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "thumbnail_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "updated_at": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "Reaction": {
        "type": "object",
        "required": [
          "emoji",
          "user_ids"
        ],
        "properties": {
          "emoji": {
            "type": "string"
          },
          "user_ids": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/String"
            }
          }
        }
      },
...
      "String": {
        "type": "string"
      },
...
      "Upload": {
        "type": "object",
        "required": [
          "id",
          "channel_id",
          "filename",
          "content_type",
          "size",
          "offset",
          "complete"
        ],
        "properties": {
          "channel_id": {
            "$ref": "#/components/schemas/String"
          },
          "complete": {
            "type": "boolean"
          },
          "content_type": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "id": {
            "$ref": "#/components/schemas/String"
          },
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "thumbnail_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "Authenticated PNG preview, at most 512 by 512. None when unavailable."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "den_session"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "sessionCookie": []
    }
  ]
}